♻️ refactor: change validatorIndex and slot types from uint256 to uint64
- Update interfaces (IYieldManager, IYieldProvider) to use uint64
- Update implementations (YieldManager, LidoStVaultYieldProvider)
- Update storage layout mapping types
- Update test and mock contracts
- Remove unnecessary type casts in LidoStVaultYieldProvider
- Add explicit casts in TestYieldManager for type compatibility

+++ contracts/contracts/_testing/mocks/yield/MockYieldProvider.sol
@@ -47,8 +47,8 @@ contract MockYieldProvider is IYieldProvider, MockYieldProviderStorageLayout {
 
   function unstakePermissionless(
     address _yieldProvider,
-    uint256 _validatorIndex,
-    uint256 _slot,
+    uint64 _validatorIndex,
+    uint64 _slot,
     bytes calldata _withdrawalParams,
     bytes calldata _withdrawalParamsProof
   ) external payable returns (uint256 maxUnstakeAmount) {
