+++ contracts/contracts/yield/interfaces/IYieldProvider.sol
@@ -108,8 +108,8 @@ interface IYieldProvider {
    */
   function unstakePermissionless(
     address _yieldProvider,
-    uint256 _validatorIndex,
-    uint256 _slot,
+    uint64 _validatorIndex,
+    uint64 _slot,
     bytes calldata _withdrawalParams,
     bytes calldata _withdrawalParamsProof
   ) external payable returns (uint256 maxUnstakeAmount);
